refactor(spellcircle): inject network execution and share scene ingestion
Accept a host-owned Boost.Asio executor in the UDP receiver. Serialize socket
operations and retire callbacks on stop or rebind without owning, running, or
joining the host context. Report asynchronous binding and receive failures.

Separate verified scene ingestion from rendering. Share document retention,
deduplication, generations, and receive-time packet rates across Qt and macOS.
Inject application runtimes explicitly and discard obsolete queued deliveries.

Validation: both applications built and displayed a live UDP scene; scoped
format/lint checks and documentation probes passed; 47 selected tests passed,
30 lifecycle/ingestion cases passed ASan/UBSan, and 14 transport cases passed
ThreadSanitizer.

+++ apps/spell-circle-canvas/src/spellcircle/qt/src/SpellCircleRenderer.cpp
@@ -18,7 +18,7 @@ void SpellCircleRenderer::synchronize(QCanvasPainterItem* item) {
 
   bool needsResolve = false;
 
-  const int modelGeneration =
+  const uint64_t modelGeneration =
       spellCircleItem->model() ? spellCircleItem->model()->generation() : 0;
   if (modelGeneration != m_knownModelGeneration) {
     m_knownModelGeneration = modelGeneration;
